Cast to unsigned long for MSVC

+++ verify/main.c
@@ -11,7 +11,7 @@ int main(int const argc, const char * const argv[])
     rom = sfc_create_rom(argv[1], strlen(argv[1]), NULL, NULL);
     if (rom == NULL)
         return 1;
-    printf("Data: %p, Size: %lu", rom->data, rom->size);
+    printf("Data: %p, Size: %lu", rom->data, (unsigned long) rom->size);
     sfc_destroy_rom(rom);
     return 0;
 }
